requirements.txt
stravalib
Flask
mysql-connector-python